chore: move legacy message to legacy folder into messages
