+++ petisco/base/domain/message/legacy/legacy_message.py
